+++ src/lib.rs
@@ -119,16 +119,6 @@ impl KeepAlive {
     }
 }
 
-/// Whether to perform a graceful shutdown or abort the connection after all requests have been handled.
-#[derive(Debug, Clone, Copy)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum ShutdownMethod {
-    /// Perform a graceful shutdown, waiting for the client to acknowledge the closure
-    Shutdown,
-    /// Abort the connection
-    Abort,
-}
-
 /// Server Configuration.
 #[derive(Debug, Clone)]
 pub struct Config<D> {
@@ -136,8 +126,6 @@ pub struct Config<D> {
     pub timeouts: Timeouts<D>,
     /// Whether to close the connection after handling a request or keeping it open to allow further requests on the same connection.
     pub connection: KeepAlive,
-    /// Whether to perform a graceful shutdown or abort the connection after all requests have been handled.
-    pub shutdown_method: ShutdownMethod,
 }
 
 impl<D> Config<D> {
@@ -147,7 +135,6 @@ impl<D> Config<D> {
         Self {
             timeouts,
             connection: KeepAlive::Close,
-            shutdown_method: ShutdownMethod::Shutdown,
         }
     }
 
@@ -164,20 +151,6 @@ impl<D> Config<D> {
     /// This is the default, but allows the configuration to be more explicit.
     pub const fn close_connection_after_response(mut self) -> Self {
         self.connection = KeepAlive::Close;
-
-        self
-    }
-
-    /// Perform a graceful shutdown after handling all requests, waiting for the client to acknowledge the closure
-    pub const fn shutdown_connection_on_close(mut self) -> Self {
-        self.shutdown_method = ShutdownMethod::Shutdown;
-
-        self
-    }
-
-    /// Abort the connection after handling all requests, without waiting for the client to acknowledge the closure
-    pub const fn abort_connection_on_close(mut self) -> Self {
-        self.shutdown_method = ShutdownMethod::Abort;
 
         self
     }
@@ -303,10 +276,7 @@ async fn serve_and_shutdown<State, T: Timer, P: routing::PathRouter<State>, S: i
     }
     .await;
 
-    let shutdown_result = match config.shutdown_method {
-        ShutdownMethod::Shutdown => socket.shutdown(&config.timeouts, &mut timer).await,
-        ShutdownMethod::Abort => socket.abort(&config.timeouts, &mut timer).await,
-    };
+    let shutdown_result = socket.shutdown(&config.timeouts, &mut timer).await;
 
     let request_count = result?;