+++ src/tests.rs
@@ -155,14 +155,6 @@ impl<TX: io::Write<Error = Infallible>, RX: io::Read<Error = Infallible>> io::So
         (&mut self.rx, &mut self.tx)
     }
 
-    async fn abort<Timer: crate::Timer>(
-        self,
-        _timeouts: &crate::Timeouts<Timer::Duration>,
-        _timer: &mut Timer,
-    ) -> Result<(), crate::Error<Self::Error>> {
-        Ok(())
-    }
-
     async fn shutdown<Timer: time::Timer>(
         self,
         _timeouts: &Timeouts<Timer::Duration>,